# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/coins":{"get":{"tags":["coin-controller"],"operationId":"getCoins","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Coin"}}}}}}},"post":{"tags":["coin-controller"],"operationId":"addCoin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Coin"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Coin"}}}}}},"delete":{"tags":["coin-controller"],"operationId":"removeCoin","parameters":[],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/coins/{id}/history":{"get":{"tags":["coin-controller"],"summary":"Kursverlauf eines Coins abrufen","operationId":"getCoinHistory","parameters":[{"name":"id","in":"path","description":"Coin ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Coin history as JSON","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CoinPriceHistory"}}}}}}}},"/api/coins/{id}/current":{"get":{"tags":["coin-controller"],"operationId":"getCoinCurrent","parameters":[{"name":"id","in":"path","description":"Coin ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/coins/watchlist":{"get":{"tags":["coin-controller"],"operationId":"getWatchlist","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Coin":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"symbol":{"type":"string"},"name":{"type":"string"},"image":{"type":"string"},"currentPrice":{"type":"number"},"marketCap":{"type":"number"},"priceChange24hPerCent":{"type":"number"},"ath":{"type":"number"}}},"CoinPriceHistory":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"symbol":{"type":"string"},"timestamp":{"type":"integer","format":"int64"},"price":{"type":"number","format":"double"},"market_cap":{"type":"number","format":"double"},"total_volume":{"type":"number","format":"double"},"marketCap":{"type":"number","format":"double","writeOnly":true},"totalVolume":{"type":"number","format":"double","writeOnly":true}}}}}}